feat: add struct ModelConfig, composed into each Provider specific config
Add in a new ModelConfig struct, as well as a trait ProviderModelConfig
which should be implemented on each Provider to propgate up model
specific configuration when used by the main agent.rs loop.

ModelConfig adds the context_limit variable and moves the name, temperature
and max_tokens configuration within it.

context_limit is set by the following order of precedence:
1. explicit context_limit
2. defaults by model name (see get_model_specific_limit)
3. a global default set to 128_000